requirements.txt
pdf2image==1.17.0
pillow==10.4.0
google-generativeai==0.8.3
python-dotenv==1.0.0
